Move error handling logic into ApplicationController base
